implement LoginModeOAuth2ActiveCondition tests
